# core/docs/swagger.yaml
definitions:
  adaptors.AdaptorMeta:
    properties:
      configSchema:
        additionalProperties: {}
        type: object
      defaultBaseUrl:
        type: string
      keyHelp:
        type: string
      name:
        type: string
      readme:
        type: string
    type: object
...
  controller.AddChannelRequest:
    properties:
      base_url:
        type: string
      configs:
        $ref: '#/definitions/model.ChannelConfigs'
      enabled_auto_balance_check:
        type: boolean
      enabled_no_permission_ban:
        type: boolean
      key:
        type: string
      max_error_rate:
        type: number
      model_mapping:
        additionalProperties:
          type: string
        type: object
      models:
        items:
          type: string
        type: array
      name:
        type: string
      priority:
        type: integer
      proxy_url:
        type: string
      sets:
        items:
          type: string
        type: array
      skip_tls_verify:
        type: boolean
      status:
        type: integer
      type:
        $ref: '#/definitions/model.ChannelType'
      warn_error_rate:
        type: number
    type: object
...
  controller.BuiltinModelConfig:
    properties:
      config:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      exclude_from_tests:
        type: boolean
      force_save_detail:
        type: boolean
      max_image_generation_count:
        type: integer
      max_video_generation_count:
        type: integer
      max_video_generation_seconds:
        type: integer
      model:
        type: string
      owner:
        $ref: '#/definitions/model.ModelOwner'
      plugin:
        additionalProperties:
          additionalProperties: {}
          type: object
        type: object
      price:
        $ref: '#/definitions/model.Price'
      request_body_storage_max_size:
        type: integer
      response_body_storage_max_size:
        type: integer
      retry_times:
        type: integer
      rpm:
        type: integer
      summary_claude_long_context:
        type: boolean
      summary_service_tier:
        type: boolean
      timeout_config:
        $ref: '#/definitions/model.TimeoutConfig'
      tpm:
        type: integer
      type:
        $ref: '#/definitions/mode.Mode'
      updated_at:
        type: string
    type: object
...
  controller.UpdateChannelStatusRequest:
    properties:
      status:
        type: integer
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  model.Channel:
    properties:
      balance:
        type: number
      balance_threshold:
        type: number
      balance_updated_at:
        type: string
      base_url:
        type: string
      channel_tests:
        items:
          $ref: '#/definitions/model.ChannelTest'
        type: array
      configs:
        $ref: '#/definitions/model.ChannelConfigs'
      created_at:
        type: string
      enabled_auto_balance_check:
        type: boolean
      enabled_no_permission_ban:
        type: boolean
      id:
        type: integer
      key:
        type: string
      last_test_error_at:
        type: string
      max_error_rate:
        type: number
      model_mapping:
        additionalProperties:
          type: string
        type: object
      models:
        items:
          type: string
        type: array
      name:
        type: string
      priority:
        type: integer
      proxy_url:
        type: string
      request_count:
        type: integer
      retry_count:
        type: integer
      sets:
        items:
          type: string
        type: array
      skip_tls_verify:
        type: boolean
      status:
        type: integer
      type:
        $ref: '#/definitions/model.ChannelType'
      used_amount:
        type: number
      warn_error_rate:
        type: number
    type: object
  model.ChannelBasicInfo:
    properties:
      id:
        type: integer
      name:
        type: string
      type:
        $ref: '#/definitions/model.ChannelType'
    type: object
  model.ChannelConfigs:
    additionalProperties: {}
    type: object
  model.ChannelTest:
    properties:
      actual_model:
        type: string
      channel_id:
        type: integer
      channel_name:
        type: string
      channel_type:
        $ref: '#/definitions/model.ChannelType'
      code:
        type: integer
      mode:
        $ref: '#/definitions/mode.Mode'
      model:
        type: string
      response:
        type: string
      success:
        type: boolean
      test_at:
        type: string
      took:
        type: number
    type: object
  model.ChannelType:
    enum:
    - 1
    - 3
    - 4
    - 12
    - 13
    - 14
    - 15
    - 16
    - 17
    - 18
    - 19
    - 20
    - 23
    - 24
    - 25
    - 26
    - 27
    - 28
    - 29
    - 30
    - 31
    - 32
    - 33
    - 34
    - 35
    - 36
    - 37
    - 40
    - 41
    - 42
    - 43
    - 44
    - 45
    - 46
    - 47
    - 48
    - 49
    - 50
    - 51
    - 52
    - 53
    - 54
    - 55
    type: integer
    x-enum-varnames:
    - ChannelTypeOpenAI
    - ChannelTypeAzure
    - ChannelTypeAzure2
    - ChannelTypeGoogleGeminiOpenAI
    - ChannelTypeBaiduV2
    - ChannelTypeAnthropic
    - ChannelTypeBaidu
    - ChannelTypeZhipu
    - ChannelTypeAli
    - ChannelTypeXunfei
    - ChannelTypeAI360
    - ChannelTypeOpenRouter
    - ChannelTypeTencent
    - ChannelTypeGoogleGemini
    - ChannelTypeMoonshot
    - ChannelTypeBaichuan
    - ChannelTypeMinimax
    - ChannelTypeMistral
    - ChannelTypeGroq
    - ChannelTypeOllama
    - ChannelTypeLingyiwanwu
    - ChannelTypeStepfun
    - ChannelTypeAWS
    - ChannelTypeCoze
    - ChannelTypeCohere
    - ChannelTypeDeepseek
    - ChannelTypeCloudflare
    - ChannelTypeDoubao
    - ChannelTypeNovita
    - ChannelTypeVertexAI
    - ChannelTypeSiliconflow
    - ChannelTypeDoubaoAudio
    - ChannelTypeXAI
    - ChannelTypeDoc2x
    - ChannelTypeJina
    - ChannelTypeTextEmbeddingsInference
    - ChannelTypeQianfan
    - ChannelTypeSangforAICP
    - ChannelTypeStreamlake
    - ChannelTypeZhipuCoding
    - ChannelTypeFake
    - ChannelTypeAntLing
    - ChannelTypeFakeError
...
  model.TimeoutConfig:
    properties:
      request_timeout:
        type: integer
      stream_request_timeout:
        type: integer
    type: object
...
info:
  contact: {}
  title: AI Proxy Swagger API
  version: "1.0"
paths:
...
  /api/channel/:
    post:
      consumes:
      - application/json
      description: Adds a new channel to the system
      parameters:
      - description: Channel information
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/controller.AddChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Add a single channel
      tags:
      - channel
  /api/channel/{id}:
    delete:
      description: Deletes a channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete a channel
      tags:
      - channel
    get:
      description: Returns detailed information about a specific channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Channel'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get a channel by ID
      tags:
      - channel
    put:
      consumes:
      - application/json
      description: Updates an existing channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated channel information
        in: body
        name: channel
        required: true
        schema:
          $ref: '#/definitions/controller.AddChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Channel'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update a channel
      tags:
      - channel
  /api/channel/{id}/{model}:
    get:
      description: Tests a single model in the channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Model name
        in: path
        name: model
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.ChannelTest'
              type: object
      security:
      - ApiKeyAuth: []
      summary: Test channel model
      tags:
      - channel
...
  /api/channel/{id}/status:
    post:
      consumes:
      - application/json
      description: Updates the status of a channel by its ID
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status information
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/controller.UpdateChannelStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Update channel status
      tags:
      - channel
...
  /api/channels/:
    get:
      description: Returns a paginated list of channels with optional filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by id
        in: query
        name: id
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by key
        in: query
        name: key
        type: string
      - description: Filter by channel type
        in: query
        name: channel_type
        type: integer
      - description: Filter by base URL
        in: query
        name: base_url
        type: string
      - description: Order by field
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        channels:
                          items:
                            $ref: '#/definitions/model.Channel'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channels with pagination
      tags:
      - channels
    post:
      consumes:
      - application/json
      description: Adds multiple channels in a batch operation
      parameters:
      - description: Channel information
        in: body
        name: channels
        required: true
        schema:
          items:
            $ref: '#/definitions/controller.AddChannelRequest'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Add multiple channels
      tags:
      - channels
  /api/channels/all:
    get:
      description: Returns a list of all channels without pagination
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Channel'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get all channels
      tags:
      - channels
...
  /api/channels/batch_delete:
    post:
      consumes:
      - application/json
      description: Deletes multiple channels by their IDs
      parameters:
      - description: Channel IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/middleware.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete multiple channels
      tags:
      - channels
  /api/channels/batch_info:
    post:
      consumes:
      - application/json
      description: Returns id, name, and type for a batch of channel IDs
      parameters:
      - description: Channel IDs
        in: body
        name: ids
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.ChannelBasicInfo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get basic info for multiple channels
      tags:
      - channels
...
  /api/channels/search:
    get:
      description: Search channels with keyword and optional filters
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Filter by id
        in: query
        name: id
        type: integer
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by key
        in: query
        name: key
        type: string
      - description: Filter by channel type
        in: query
        name: channel_type
        type: integer
      - description: Filter by base URL
        in: query
        name: base_url
        type: string
      - description: Order by field
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    allOf:
                    - {}
                    - properties:
                        channels:
                          items:
                            $ref: '#/definitions/model.Channel'
                          type: array
                        total:
                          type: integer
                      type: object
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Search channels
      tags:
      - channels
...
  /api/channels/type_metas:
    get:
      description: Returns metadata for all channel types
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    $ref: '#/definitions/adaptors.AdaptorMeta'
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel type metadata
      tags:
      - channels
...
  /api/models/builtin/channel:
    get:
      description: Returns a list of channel builtin models
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  additionalProperties:
                    items:
                      $ref: '#/definitions/controller.BuiltinModelConfig'
                    type: array
                  type: object
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel builtin models
      tags:
      - model
  /api/models/builtin/channel/{type}:
    get:
      description: Returns a list of channel builtin models by type
      parameters:
      - description: Channel type
        enum:
        - 1
        - 3
        - 4
        - 12
        - 13
        - 14
        - 15
        - 16
        - 17
        - 18
        - 19
        - 20
        - 23
        - 24
        - 25
        - 26
        - 27
        - 28
        - 29
        - 30
        - 31
        - 32
        - 33
        - 34
        - 35
        - 36
        - 37
        - 40
        - 41
        - 42
        - 43
        - 44
        - 45
        - 46
        - 47
        - 48
        - 49
        - 50
        - 51
        - 52
        - 53
        - 54
        - 55
        in: path
        name: type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controller.BuiltinModelConfig'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get channel builtin models by type
      tags:
      - model
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"